:hammer: MemberService 충돌 해결
