+++ src/main/java/com/example/scheduo/global/config/security/SecurityConfig.java
@@ -44,7 +44,7 @@ public class SecurityConfig {
 			.httpBasic(AbstractHttpConfigurer::disable)
 			.sessionManagement((session) -> session
 				.sessionCreationPolicy(SessionCreationPolicy.STATELESS))
-			.authorizeHttpRequests(request -> request.anyRequest().authenticated())
+			.authorizeHttpRequests(request -> request.anyRequest().permitAll())
 			.oauth2Login(oauth2 -> oauth2
 				.userInfoEndpoint(userInfo -> userInfo.userService(customOAuth2UserService))
 				.successHandler(customOAuth2SuccessHandler)
